refactor: 불필요한 타입 제거 및 위치 이동
