+++ src/app/library/detail/[id]/_components/SideButtonGroup.tsx
@@ -1,7 +1,7 @@
 'use client';
 import { ChevronLeft, Heart } from 'lucide-react';
 import { useParams, useRouter } from 'next/navigation';
-import { useAuth } from '@/providers/auth-provider/AuthProvider.client';
+import { useAuth } from '@/providers/auth/AuthProvider.client';
 
 import useLikeStory from '@/hooks/api/library/useLikeStory';
 
@@ -35,7 +35,6 @@ const SideButtonGroup = () => {
         className="flex-center md:border-write-main h-10 w-10 rounded-full border-0 bg-white md:h-12 md:w-12 md:border"
         onClick={() => router.back()}
       >
-
         <ChevronLeft className="text-write-main h-8 w-8" aria-hidden />
       </button>
 
@@ -55,7 +54,6 @@ const SideButtonGroup = () => {
         <span className="text-write-main h-3 text-xs">
           {likeCount ? (likeCount > 999 ? '999+' : likeCount) : 0}
         </span>
-
       </button>
     </div>
   );
