+++ src/app/mypage/_components/my-social-list/MySocialList.tsx
@@ -3,7 +3,7 @@
 import Observer from '@/components/common/Observer/Observer';
 import { useEffect, useState } from 'react';
 import { useMySocialList } from '@/hooks/api/mypage/useMySocialList';
-import { useAuth } from '@/providers/auth-provider/AuthProvider.client';
+import { useAuth } from '@/providers/auth/AuthProvider.client';
 import { TabType } from './type';
 import TabMenu from './TabMenu';
 import MySocialListCard from './MySocialListCard';
