+++ src/app/social/detail/[storyId]/page.tsx
@@ -4,7 +4,7 @@ import { dehydrate, HydrationBoundary } from '@tanstack/react-query';
 import { SocialDetailPageParams } from '@/app/social/detail/[storyId]/type';
 import SocialOverView from '@/app/social/detail/[storyId]/_components/SocialOverView';
 import StorySummary from '@/app/social/detail/[storyId]/_components/StorySummary';
-import getMyInfoOnServer from '@/providers/auth-provider/getMyInfoOnServer';
+import getMyInfoOnServer from '@/providers/auth/getMyInfoOnServer';
 import { checkStoryExists, getStory } from '@/api/stories/api';
 import { getStoryCollaborators } from '@/api/story-collaborators/api';
 import { notFound } from 'next/navigation';
