--- src/types/teamUserRole.ts
@@ -1,7 +1,0 @@
-export const TEAM_USER_ROLE = {
-  LEADER: 'LEADER',
-  MEMBER: 'MEMBER',
-  GUEST: 'GUEST',
-} as const;
-
-export type TeamUserRole = (typeof TEAM_USER_ROLE)[keyof typeof TEAM_USER_ROLE];
