--- src/types/textEditor.ts
@@ -1,4 +1,0 @@
-export interface TextEditorRef {
-  getHTML: () => string;
-  getText: () => string;
-}
